+++ src/test/java/com/jnape/palatable/lambda/functions/IOTest.java
@@ -1,0 +1,32 @@
+package com.jnape.palatable.lambda.functions;
+
+import com.jnape.palatable.traitor.annotations.TestTraits;
+import com.jnape.palatable.traitor.runners.Traits;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import testsupport.EqualityAwareIO;
+import testsupport.traits.ApplicativeLaws;
+import testsupport.traits.FunctorLaws;
+import testsupport.traits.MonadLaws;
+
+import static com.jnape.palatable.lambda.adt.Unit.UNIT;
+import static com.jnape.palatable.lambda.functions.Fn0.fn0;
+import static com.jnape.palatable.lambda.functions.IO.io;
+import static org.junit.Assert.assertEquals;
+
+@RunWith(Traits.class)
+public class IOTest {
+
+    @TestTraits({FunctorLaws.class, ApplicativeLaws.class, MonadLaws.class})
+    public IO<Integer> testSubject() {
+        return new EqualityAwareIO<>(io(1));
+    }
+
+    @Test
+    public void staticFactoryMethods() {
+        assertEquals((Integer) 1, io(1).unsafePerformIO());
+        assertEquals((Integer) 1, io(() -> 1).unsafePerformIO());
+        assertEquals((Integer) 1, io(fn0(() -> 1)).unsafePerformIO());
+        assertEquals(UNIT, io(() -> {}).unsafePerformIO());
+    }
+}